+++ cdn/distribution/locals.tf
@@ -1,9 +1,10 @@
 module "name_map" {
-  source                = "../../name_map"
-  name_infix_default    = var.domain_name_infix_default
-  name_map              = local.l0_map
-  name_regex_allow_list = ["."]
-  std_map               = var.std_map
+  source                          = "../../name_map"
+  name_include_app_fields_default = var.name_include_app_fields_default
+  name_infix_default              = var.name_infix_default
+  name_map                        = local.l0_map
+  name_regex_allow_list           = ["."]
+  std_map                         = var.std_map
 }
 
 locals {