+++ iam/user_group/locals.tf
@@ -61,7 +61,7 @@ locals {
     user = {
       for k, v in local.user_map : k => merge(
         {
-          for k_user, v_user in v : k_user => v_user if !contains(["role_policy_inline_json_map"], k_user)
+          for k_user, v_user in v : k_user => v_user if !contains([], k_user)
         },
         {
           arn = aws_iam_user.this_user[k].arn